Consolidate sort/filter helpers and host normalization
